+++ src/ope/dfiv/nn_structure/__init__.py
@@ -3,15 +3,14 @@ from acme.tf import utils as tf2_utils
 import sonnet as snt
 from typing import Tuple
 from .cartpole_swingup import make_value_func_cartpole
-from .bsuite_cartpole_swingup import make_value_func_bsuite_cartpole
-
+from .bsuite_network import make_value_func_bsuite
 
 def make_ope_networks(task_id: str, environment_spec: EnvironmentSpec) -> Tuple[snt.Module, snt.Module]:
 
-    if task_id == "cartpole_swingup" or task_id == "dm_control_cartpole_swingup":
+    if task_id == "dm_control_cartpole_swingup":
         value_func, instrumental_feature = make_value_func_cartpole()
-    elif task_id == "bsuite_cartplot_swingup":
-        value_func, instrumental_feature = make_value_func_bsuite_cartpole()
+    elif task_id.startswith("bsuite"):
+        value_func, instrumental_feature = make_value_func_bsuite(environment_spec)
     else:
         raise ValueError(f"task id {task_id} not known")
