--- src/ope/dfiv/nn_structure/bsuite_cartpole_swingup.py
@@ -1,48 +1,0 @@
-from typing import Tuple
-import sonnet as snt
-import tensorflow as tf
-from acme.tf import utils as tf2_utils
-from acme.tf import networks
-from acme.specs import EnvironmentSpec
-
-
-class InstrumentalFeature(snt.Module):
-
-    def __init__(self):
-        super(InstrumentalFeature, self).__init__()
-        self._net = snt.Sequential([networks.LayerNormMLP([1024,1024,512], activate_final=True)])
-
-    def __call__(self, obs, action):
-        obs_aug = tf.squeeze(obs, axis=1)
-        action_aug = tf.one_hot(action, depth=3)
-        data = tf.concat([obs_aug, action_aug], axis=1)
-        return self._net(data)
-
-class ValueFeature(snt.Module):
-
-    def __init__(self):
-        super(ValueFeature, self).__init__()
-        self._net = snt.Sequential([snt.Sequential([networks.LayerNormMLP([512, 512, 512], activate_final=True)])])
-
-    def __call__(self, obs, action):
-        obs_aug = tf.squeeze(obs, axis=1)
-        action_aug = tf.one_hot(action, depth=3)
-        data = tf.concat([obs_aug, action_aug], axis=1)
-        return self._net(data)
-
-
-class ValueFunction(snt.Module):
-
-    def __init__(self):
-        super(ValueFunction, self).__init__()
-        self._feature = ValueFeature()
-        self._weight = tf.random.uniform((512, 1))
-
-    def __call__(self, obs, action):
-        return tf.matmul(self._feature(obs, action), self._weight)
-
-
-def make_value_func_bsuite_cartpole() -> Tuple[snt.Module, snt.Module]:
-    value_function = ValueFunction()
-    instrumental_feature = InstrumentalFeature()
-    return value_function, instrumental_feature
